+++ tests/test_field.py
@@ -64,7 +64,7 @@ def pre_mixed():
 
     u = fem.Field(r, dim=3)
     p = fem.Field(r)
-    J = fem.Field(r, values=1)
+    J = fem.Field(r, values=1.0)
 
     f = fem.FieldContainer([u, p, J])
     g = fem.FieldsMixed(fem.RegionHexahedron(m), n=3, disconnect=True)
@@ -98,7 +98,7 @@ def pre_axi_mixed():
 
     u = fem.FieldAxisymmetric(r, dim=2)
     p = fem.Field(r)
-    J = fem.Field(r, values=1)
+    J = fem.Field(r, values=1.0)
 
     f = fem.FieldContainer((u, p, J))